Remaking the whole site to better adhere to best practices
The original project had become messy and there were several areas that
needed to be reworked. This happened because we tried to implement the
whole site with dummy components and then integrating the working
components as they became ready. Instead of fixing the old code, it
seemed cleaner to start a new project and build upon it as the backend
functionalities became ready. i.e. Only when an API is ready will the
necessary HTML and JS be created. This check-in includes:
-Created a new SPA based on AngularJS seed project
-Added basic navigation and header